test-requirements.txt
nose
libvirt-python
flake8
